move timestamps.py to utils
