+++ tests/adapter/dbt/tests/adapter/utils/test_timestamps.py
