+++ tests/functional/timestamps/test_timestamps.py
@@ -1,4 +1,4 @@
-from tests.adapter.dbt.tests.adapter.basic.test_timestamps import TestCurrentTimestamps
+from tests.adapter.dbt.tests.adapter.utils.test_timestamps import TestCurrentTimestamps
 
 
 class TestCurrentTimestampsCore(TestCurrentTimestamps):
